feat: nested package list
